Tests/LibWeb/Text/expected/css/white-space-trim.txt

none
discard-inner
discard-after discard-inner
